move tests into their own crate, remove the shared crate by moving its necessary code into pg_search, and cleanup dependencies
