+++ pg_search-tests/tests/cleanup.rs
