+++ pg_search-tests/tests/custom_scan.rs
