+++ pg_search-tests/tests/find_var_relation.rs
@@ -17,9 +17,9 @@
 
 mod fixtures;
 
+use fixtures::db::Query;
 use fixtures::*;
 use rstest::*;
-use shared::fixtures::db::Query;
 use sqlx::PgConnection;
 
 #[rstest]
