+++ pg_search-tests/tests/fixtures/tables/duckdb_types.rs
