+++ pg_search-tests/tests/fixtures/tables/icu_arabic_posts.rs
