+++ pg_search-tests/tests/fixtures/tables/mod.rs
@@ -22,7 +22,6 @@ mod icu_arabic_posts;
 mod icu_czech_posts;
 mod icu_greek_posts;
 mod nyc_trips;
-mod research_project_arrays;
 mod simple_products;
 mod user_session_logs;
 
@@ -33,6 +32,5 @@ pub use icu_arabic_posts::*;
 pub use icu_czech_posts::*;
 pub use icu_greek_posts::*;
 pub use nyc_trips::*;
-pub use research_project_arrays::*;
 pub use simple_products::*;
 pub use user_session_logs::*;
