+++ pg_search-tests/tests/fixtures/tables/simple_products.rs
