+++ pg_search-tests/tests/fixtures/utils.rs
@@ -14,7 +14,7 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
-
+#![allow(dead_code)]
 use super::db::*;
 
 use sqlx::PgConnection;
