+++ pg_search/src/trace.rs
